requirements.txt
alembic==1.6.3
aniso8601==9.0.1
astroid==2.5.6
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
Flask-JWT-Extended==4.2.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
mccabe==0.6.1
passlib==1.7.4
pkg-resources==0.0.0
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.8.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
Werkzeug==2.0.1
wrapt==1.12.1
